refactor(web): update session API path to internal endpoint
Updated the session API path in `getSession.ts` to use the internal
endpoint `/api/internal/auth/local/me` for better alignment with
the new internal API structure.

BREAKING CHANGE: The session API path has been changed to an internal
endpoint, which may require updates in dependent services or clients.
