--- apps/api/src/modules/auth/index.ts
@@ -1,2 +1,0 @@
-export * from './auth.module';
-export * from './local';
