--- apps/api/src/modules/auth/local/dtos/index.ts
@@ -1,2 +1,0 @@
-export * from './login.dto';
-export * from './register.dto';
