--- apps/api/src/modules/auth/local/index.ts
@@ -1,4 +1,0 @@
-export * from './local.controller';
-export * from './local.service';
-export * from './local.strategy';
-export * from './local-session-serializer';
