--- apps/api/src/modules/index.ts
@@ -1,2 +1,0 @@
-export * from './auth/auth.module';
-export * from './users/users.module';
