+++ apps/api/src/modules/internal/auth/auth.module.ts
@@ -1,7 +1,10 @@
 import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { UsersModule } from '../users';
-import { LocalAuthService, LocalAuthController, LocalStrategy, LocalAuthSessionSerializer } from './local';
+import { LocalAuthController } from './controllers';
+import { LocalAuthService } from './services';
+import { LocalAuthSessionSerializer } from './session-serializers/';
+import { LocalStrategy } from './strategies/';
 
 @Module({
   imports: [UsersModule, PassportModule],
