+++ apps/api/src/modules/internal/auth/controllers/index.ts
@@ -1,0 +1,1 @@
+export * from './local.controller';
