+++ apps/api/src/modules/internal/auth/dtos/index.ts
@@ -1,0 +1,2 @@
+export * from './local-register.dto';
+export * from './local-login.dto';
