+++ apps/api/src/modules/internal/auth/guards/index.ts
@@ -1,0 +1,1 @@
+export * from './local.guard';
