+++ apps/api/src/modules/internal/auth/index.ts
@@ -1,0 +1,7 @@
+export * from './controllers';
+export * from './dtos';
+export * from './guards';
+export * from './services';
+export * from './session-serializers';
+export * from './strategies';
+export * from './auth.module';
