+++ apps/api/src/modules/internal/auth/services/index.ts
@@ -1,0 +1,1 @@
+export * from './local.service';
