+++ apps/api/src/modules/internal/auth/session-serializers/index.ts
@@ -1,0 +1,1 @@
+export * from './local-session-serializer';
