+++ apps/api/src/modules/internal/auth/strategies/index.ts
@@ -1,0 +1,1 @@
+export * from './local.strategy';
