+++ apps/api/src/modules/internal/users/controllers/index.ts
@@ -1,0 +1,1 @@
+export * from './users.controller';
