+++ apps/api/src/modules/internal/users/index.ts
@@ -1,0 +1,6 @@
+export * from './controllers';
+export * from './dtos';
+export * from './repositories';
+export * from './schemas';
+export * from './services';
+export * from './users.module';
