+++ apps/api/src/modules/internal/users/repositories/index.ts
@@ -1,0 +1,1 @@
+export * from './users.repository';
