+++ apps/api/src/modules/internal/users/services/index.ts
@@ -1,0 +1,1 @@
+export * from './users.service';
