+++ apps/api/src/modules/internal/users/users.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { User, UserSchema } from './schemas';
-import { UsersService } from './users.service';
-import { UsersController } from './users.controller';
-import { UsersRepository } from './users.repository';
+import { UsersService } from './services';
+import { UsersController } from './controllers';
+import { UsersRepository } from './repositories';
 
 @Module({
   imports: [MongooseModule.forFeature([{ name: User.name, schema: UserSchema }], 'main')],
