--- apps/api/src/modules/users/index.ts
@@ -1,4 +1,0 @@
-export * from './dtos';
-export * from './schemas';
-export * from './users.module';
-export * from './users.service';
