+++ apps/api/src/types.ts
@@ -1,0 +1,7 @@
+import 'express';
+
+declare module 'express' {
+  export interface Request {
+    // type here
+  }
+}
